Rename nvidia shadowplay directory
Recordings could come from anywhere (e.g. OBS, AMD Relive).
