+++ auto-handbrake-cfr/src/main/resources/log4j2.xml
