+++ auto-handbrake-cfr/src/test/resources/Big_Buck_Bunny_360_10s_2MB.mp4
