--- nvidia-shadowplay/build.gradle
@@ -1,13 +1,0 @@
-String inputDirectory = project.getProperties().getOrDefault("inputDirectory", "")
-String outputDirectory = project.getProperties().getOrDefault("outputDirectory", "")
-String archiveDirectory = project.getProperties().getOrDefault("archiveDirectory", "")
-
-task run(type: JavaExec) {
-  classpath = sourceSets.main.runtimeClasspath
-  main = 'com.willmolloy.handbrake.nvidia.shadowplay.App'
-  args = [inputDirectory, outputDirectory, archiveDirectory]
-}
-
-dependencies {
-  implementation project(":auto-handbrake-core")
-}
